fix in salvataggio prenotazione; fix in salvataggio ordine chiuso

+++ code/app/Http/Controllers/OrdersController.php
@@ -207,37 +207,39 @@ class OrdersController extends Controller
         $availables = $request->input('product_max_available', []);
         $products = $request->input('productid');
 
-        for ($i = 0; $i < count($products); ++$i) {
-            $id = $products[$i];
+        if (count($prices)) {
+            for ($i = 0; $i < count($products); ++$i) {
+                $id = $products[$i];
 
-            foreach ($enabled as $en) {
-                if ($en == $id) {
-                    $new_products[] = $id;
-                    break;
+                foreach ($enabled as $en) {
+                    if ($en == $id) {
+                        $new_products[] = $id;
+                        break;
+                    }
+                }
+
+                $prod = Product::find($id);
+                if ($prod->price != $prices[$i] || $prod->transport != $transports[$i] || $prod->max_available != $availables[$i]) {
+                    $prod->price = $prices[$i];
+                    $prod->transport = $transports[$i];
+                    $prod->max_available = $availables[$i];
+                    $prod->save();
                 }
             }
 
-            $prod = Product::find($id);
-            if ($prod->price != $prices[$i] || $prod->transport != $transports[$i] || $prod->max_available != $availables[$i]) {
-                $prod->price = $prices[$i];
-                $prod->transport = $transports[$i];
-                $prod->max_available = $availables[$i];
-                $prod->save();
+            /*
+                Se vengono rimossi dei prodotti dall'ordine, ne elimino tutte le
+                relative prenotazioni sinora avvenute
+            */
+            $removed_products = $order->products()->whereNotIn('id', $new_products)->pluck('id')->toArray();
+            foreach($order->bookings as $booking) {
+                $booking->products()->whereIn('product_id', $removed_products)->delete();
+                if ($booking->products->isEmpty())
+                    $booking->delete();
             }
-        }
 
-        /*
-            Se vengono rimossi dei prodotti dall'ordine, ne elimino tutte le
-            relative prenotazioni sinora avvenute
-        */
-        $removed_products = $order->products()->whereNotIn('id', $new_products)->pluck('id')->toArray();
-        foreach($order->bookings as $booking) {
-            $booking->products()->whereIn('product_id', $removed_products)->delete();
-            if ($booking->products->isEmpty())
-                $booking->delete();
+            $order->products()->sync($new_products);
         }
-
-        $order->products()->sync($new_products);
 
         $discounted = $request->input('discounted', []);
         foreach ($order->products as $product) {
